feat(#683): improved validation

+++ packages/epics/src/people/hooks/validation.ts
@@ -16,8 +16,8 @@ export const activateSpacesSchema = z.object({
   spaces: z
     .array(
       z.object({
-        spaceId: z.number().min(1, 'Space is required'),
-        months: z.number().min(1, 'Months must be at least 1'),
+        spaceId: z.number().int().min(1, 'Space is required'),
+        months: z.number().int().min(1, 'Months must be at least 1'),
       }),
     )
     .min(1, 'At least one space must be added'),
